Creating router request and initial update user

+++ src/slices/user/controllers/deleteUser/deleteUserController.ts
@@ -1,3 +1,4 @@
+import { UnauthorizedError } from "@/application/errors";
 import {
   badRequest,
   HttpRequest,
@@ -19,6 +20,16 @@ export class DeleteUserController extends Controller {
     const errors = this.validation.validate(httpRequest?.query);
     if (errors?.length > 0) {
       return badRequest(errors);
+    }
+    if (httpRequest?.userLogged?.role === "admin") {
+      const userDeleted = await this.deleteUser({
+        fields: { ...httpRequest?.query },
+        options: {},
+      });
+      return success(userDeleted);
+    }
+    if (httpRequest?.userId !== httpRequest?.query?._id) {
+      return badRequest(UnauthorizedError);
     }
     const userDeleteed = await this.deleteUser({
       fields: { ...httpRequest?.query, createdById: httpRequest?.userId },
